% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_processing_env.R
\docType{data}
\name{create.processing.env}
\alias{create.processing.env}
\alias{run.env}
\title{Create processing environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
run.env
}
\arguments{
\item{base.dir}{character; path to working directory. It is helpful to make this explicit (rather than using '.') for generating SGE_Array commands.}

\item{raw.seq.dirs}{character; path to raw FASTQs directory \strong{or} a \emph{named} vector of paths to raw FASTQs directories. If a vector, \code{symlink.raw.fastqs} will preppend name of path to sample names to keep them distinct.}

\item{interactive}{logical; TRUE = run command interactively, no swarm computing possible, FALSE = generate SGE_Batch or SGE_Array commands to submit manually. Default is FALSE}

\item{job.queue}{character; the queue to that jobs will be sent to, for generating SGE commands. Default is NULL.}

\item{jobs.dir}{character; path to write command files to for submission to the scheduler from the appropriate machine, e.g., "/home/micro/stagamak/Jobs". Default is ".".}

\item{max.cores}{integer; maximum cores to take advantage of (when utilizing swarm processing this number will be divided by max.concurrent jobs to determine the per-job number of cores). Default is 1.}

\item{max.concurrent.jobs}{integer; maximum concurrent jobs to run. If \code{interactive} == TRUE, this will be forced to 1. Default is 1.}

\item{link.dir}{character; path to directory wherein you want to make symlinks to raw FASTQs. If set to NULL, will use the \code{base.dir}. Default is NULL.}

\item{store.dir}{character; path to directory where all output will be stored. If set to NULL, will use the \code{base.dir}. Default is NULL.}

\item{temp.dir}{character; path to a directory where output will first be written before moving to the permanent output directory, if NULL, output is written directly to \code{store.dir}. Default is NULL.}

\item{max.memory}{character; maximum memory to set for jobs submitted to SGE_Batch or SGE_Array. Must take the form of e.g. "10G" or "50G" Setting to NULL means this parameter is not specified in the command. Default is NULL.}

\item{qsub.options}{character; further qsub arguments you want included in the SGE_Batch or SGE_Array commands, e.g., get email at end of job with "'-m ae -M yourname@example.com'". This argument must be in double and single quotes like example to be parsed correctly. Setting to NULL means this parameter i snot specified in the command. Default is NULL.}

\item{save.env.dir}{character; if not NULL, will create and save this environment in the provided directory (if directory does not already exist, it will be created in \code{base.dir}) with the date and time of creation in the file name. Default is NULL.}

\item{samples}{character; a vector of sample names. Can be left NULL and \code{symlink.raw.fastqs} will grab sample names from the raw fastq file names. Default NULL.}

\item{...}{additional variables you want to set here}
}
\description{
This function creates a new environment called \code{run.env} to store all user-set variables with the option to save as an RDA file for easy reading-in to a future session.
}
\seealso{
\code{\link{assign}}, \code{\link{new.env}}, \code{\link{call_match}}, \code{\link{symlink.raw.fastqs}}
}
\keyword{datasets}
