% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_commands.R
\name{execute.commands}
\alias{execute.commands}
\title{Execute commands}
\usage{
execute.commands(commands)
}
\arguments{
\item{commands}{character; vector of shell commands.}
}
\description{
This function takes a character vector of shell commands and uses \code{system} to execute them
}
\seealso{
\code{\link{system}}
}
