% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_full_commands.R
\name{generate.full.commands}
\alias{generate.full.commands}
\title{Generate full commands}
\usage{
generate.full.commands(
  input.dir = NULL,
  paired = TRUE,
  tmp.dir = NULL,
  output.dir = ".",
  zip.output = TRUE,
  split = "--",
  sample.field = 1,
  write.to = NULL,
  ...
)
}
\arguments{
\item{input.dir}{character; path to directory containing input files. Default is NULL.}

\item{paired}{logical; are their paired R1 and R2 files that should be analyzed concurrently? Default is TRUE.}

\item{tmp.dir}{character; path to temporary directory for direct output. After files are written here, they will be moved to \code{output.dir}. Default is NULL.}

\item{output.dir}{character; path to storage directory for output. If no \code{tmp.dir} output will be written here, otherwise it will be written to \code{tmp.dir} first and then moved here. Default is '.'.}

\item{zip.output}{logical; whether to gzip final output (occurs before moving files if \code{tmp.dir} is set). Default is TRUE.}

\item{split}{character; character(s) to split the sample name from the read ID in link name, e.g. sample01-R1.fastq.gz OR sample01--R1.fastq.gz. Default is '--'".}

\item{sample.field}{integer; field (after cutting) in which to find sample names. Default is 1.}

\item{write.to}{character; file name to write commands to, if NULL, only returns commands as character vector. Default is NULL}

\item{...}{other commands to pass to appropriate tool. Names must match short or long version found in that tool's help page.}
}
\description{
This function generates a single-line tool command with the supplied parameters for each file in the supplied directory as well as appending other commands, such as gzipping and moving results files.
}
\seealso{
\code{\link{system}}, \code{\link{list2}}, \code{\link{generate.tool.command}}
}
