% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tool_command.R
\name{generate.tool.commmand}
\alias{generate.tool.commmand}
\title{Generate tool command}
\usage{
generate.tool.commmand(tool, ...)
}
\arguments{
\item{tool}{character; name of the appropriate tool, e.g. "kneaddata" or "humann".}

\item{...}{other commands to pass to appropriate tool. Names must match short or long version found in that tool's help page.}
}
\description{
This function generates a character string of a command given a tool and its arguments.
}
\seealso{
\code{\link{system}}, \code{\link{list2}}
}
