% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_chocophlan.R
\name{install.chocophlan}
\alias{install.chocophlan}
\title{Install Chocophlan Database}
\usage{
install.chocophlan(location)
}
\arguments{
\item{location}{character; path to install location.}
}
\description{
This function installs the chocophlan database for using with Humann3.
}
\seealso{
\code{\link{system}}, \code{\link{list2}}
}
