% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_metaphlan2_tables.R
\name{join.metaphlan2.tables}
\alias{join.metaphlan2.tables}
\title{Join single-sample metaphlan2 tables}
\usage{
join.metaphlan2.tables(
  match.pattern = "metaphlan_bugs_list",
  input.dir,
  output.file
)
}
\arguments{
\item{match.pattern}{character; pattern to match desired files and directories. Default is "metaphlan_bugs_list"}

\item{input.dir}{character; path to input directory.}

\item{output.file}{character; path and name of output file.}
}
\description{
This function takes multiple humann metaphlan2 bug list files and combines them into a single output file.
}
