% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_unpacked_pathways_tables.R
\name{join.unpacked.pathways.tables}
\alias{join.unpacked.pathways.tables}
\title{Join single-sample unpacked pathways tables}
\usage{
join.unpacked.pathways.tables(
  match.pattern = "unpacked_pathways_relab",
  input.dir,
  output.file
)
}
\arguments{
\item{match.pattern}{character; pattern to match desired files and directories. Default is "metaphlan_bugs_list"}

\item{input.dir}{character; path to input directory.}

\item{output.file}{character; path and name of output file.}
}
\description{
This function takes multiple humann unpacked pathwaysfiles and combines them into a single output file.
}
