% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tool_help.R
\name{print.tool.help}
\alias{print.tool.help}
\title{Print tool help}
\usage{
\method{print}{tool.help}(tool)
}
\arguments{
\item{tool}{character; name of the appropriate tool, e.g. "kneaddata" or "humann".}
}
\description{
This function prints the help page of the provided tool to reference supplying the proper arguments to it.
}
\seealso{
\code{\link{system}}, \code{\link{generate.tool.command}}
}
