% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlink_raw_fastqs.R
\name{symlink.raw.fastqs}
\alias{symlink.raw.fastqs}
\title{Create symlinks to fastq(.gz) files}
\usage{
symlink.raw.fastqs(
  fastq.dirs,
  delim,
  sample.field,
  pattern = "lane.*fastq\\\\.gz",
  split = "--",
  replacements = NULL,
  save.sample.names = TRUE
)
}
\arguments{
\item{fastq.dirs}{character; path to raw FASTQs directory \strong{or} a \emph{named} vector of paths to raw FASTQs directories. If a vector, function will preppend name of path to sample names to keep them distinct.}

\item{delim}{character; delimiter for finding sample names in fastq file names.}

\item{sample.field}{integer; field (after cutting) in which to find sample names.}

\item{pattern}{character; matching pattern for files you want to link. Default is 'lane.*fastq\\.gz'.}

\item{split}{character; character(s) to split the sample name from the read ID in link name, e.g. sample01-R1.fastq.gz OR sample01--R1.fastq.gz. Default is '--'".}

\item{replacements}{list; a named list of patterns and replacements for file names, e.g. =list(to.replace = c("A", "B"), replace.by = c("a", "b")). Default is NULL.}

\item{save.sample.names}{logical, whether to grab sample names from file names (TRUE) let user supply sample names directly (FALSE). Default is TRUE.}
}
\description{
This function creates symlinks in the current directory to fastq OR fastq.gz files in the provided directory.
}
\seealso{
\code{\link{system}}, \code{\link{str_split}}, \code{\link{list.files}}, \code{\link{create.processing.env}}
}
