% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlinkFastqs.R
\name{symlinkFastqs}
\alias{symlinkFastqs}
\title{Create symlinks to fastq(.gz) files}
\usage{
symlinkFastqs(
  fastq.dir,
  delim,
  sample.field,
  pattern = "lane.*fastq\\\\.gz",
  split = "--",
  replacements = NULL
)
}
\arguments{
\item{fastq.dir}{character; path to directory containing raw FASTQ files.}

\item{delim}{character; delimiter for finding sample names in fastq file names.}

\item{sample.field}{integer; field (after cutting) in which to find sample names.}

\item{pattern}{character; matching pattern for files you want to link. Default is 'lane.*fastq\\.gz'.}

\item{split}{character; character(s) to split the sample name from the read ID in link name, e.g. sample01-R1.fastq.gz OR sample01--R1.fastq.gz. Default is '--'".}

\item{replacements}{list; a named list of patterns and replacements for file names, e.g. =list(to.replace = c("A", "B"), replace.by = c("a", "b")). Default is NULL.}
}
\description{
This function creates symlinks in the current directory to fastq OR fastq.gz files in the provided directory.
}
\seealso{
\code{\link{system}}, \code{\link{str_split}}, \code{\link{list.files}}
}
