% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{tgz.directories}
\alias{tgz.directories}
\title{Helper functions}
\usage{
tgz.directories(location, match.pattern = NULL)
}
\arguments{
\item{location}{character; path to directory with files/directories you want to manipulate.}

\item{match.pattern}{character; pattern to match desired files and directories. Default NULL}

\item{move.from}{character; path to directory you want to move files/directories from.}

\item{move.to}{character; path to directory you want to move files/directories to.}
}
\description{
Functions for tarring, zipping, and moving files.
}
\seealso{
\code{\link{system}}, \code{\link{generate.full.commands}}
}
