% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper.functions}
\alias{helper.functions}
\alias{show.tool.help}
\alias{print.tool.help}
\alias{tgz.directories}
\alias{remove.directories}
\alias{gzip.files}
\alias{gunzip.files}
\alias{move.files}
\title{Helper functions}
\usage{
show.tool.help(tool)

tgz.directories(location, match.pattern = NULL)

remove.directories(location, match.pattern = NULL)

gzip.files(location, match.pattern = NULL)

gunzip.files(location, match.pattern = NULL)

move.files(move.from, move.to, match.pattern = NULL)
}
\arguments{
\item{tool}{character; name of the appropriate tool, e.g. "kneaddata" or "humann".}

\item{location}{character; path to directory with files/directories you want to manipulate.}

\item{match.pattern}{character; pattern to match desired files and directories. Default NULL}

\item{move.from}{character; path to directory you want to move files/directories from.}

\item{move.to}{character; path to directory you want to move files/directories to.}
}
\description{
Functions for printing help, tarring, zipping, and moving files.
}
\seealso{
\code{\link{system}}, \code{\link{generate.full.commands}}, \code{\link{generate.tool.commands}}
}
