% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{print.tool.help}
\alias{print.tool.help}
\title{Helper functions}
\usage{
\method{print}{tool.help}(tool)
}
\arguments{
\item{tool}{character; name of the appropriate tool, e.g. "kneaddata" or "humann".}

\item{match.pattern}{character; pattern to match desired files and directories. Default NULL}

\item{location}{character; path to directory with files/directories you want to manipulate.}

\item{move.from}{character; path to directory you want to move files/directories from.}

\item{move.to}{character; path to directory you want to move files/directories to.}
}
\description{
Functions for tarring, zipping, and moving files.
}
\seealso{
\code{\link{system}}, \code{\link{generate.full.commands}}
}
